requirements.txt
SQLAlchemy==2.0.27
fastapi==0.110.0
uvicorn==0.27.1
openai==1.12.0
python-dotenv==1.0.1
